refactor: move partition to its own header file
